OAUserFacet/core/api/serializers.py
from rest_framework import serializers
from .models import Survey, UserResponse, Question

class SurveySerializer(serializers.ModelSerializer):
    class Meta:
        model = Survey
        fields = '__all__'

class UserResponseSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserResponse
        fields = '__all__'

class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = ['survey'] + [f'question{i}_answer' for i in range(1, 21)]

